% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.glmcat}
\alias{anova.glmcat}
\title{Variance-Covariance Matrix for a Fitted glmcat Model Object}
\usage{
\method{anova}{glmcat}(object, object2, type, ...)
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted \code{glmcat} model object.
}
