% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{coef.glmcat}
\alias{coef.glmcat}
\title{Model coefficients of a fitted \code{glmcat} model object}
\usage{
\method{coef}{glmcat}(object, na.rm = FALSE, ...)
}
\arguments{
\item{object}{an fitted object of class \code{glmcat}.}

\item{na.rm}{TRUE for NA coefficients to be removed, default is FALSE.}

\item{...}{additional arguments affecting the \code{coef} method.}
}
\description{
Returns the coefficient estimates of the fitted \code{glmcat} model object.
}
