% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{confint.glmcat}
\alias{confint.glmcat}
\title{Confidence intervals for parameters of a fitted \code{glmcat} model object}
\usage{
\method{confint}{glmcat}(object, level, ...)
}
\arguments{
\item{object}{an fitted object of class \code{glmcat}.}

\item{level}{the confidence level.}
}
\description{
Computes confidence intervals from a fitted \code{glmcat} model object for all the parameters.
}
