% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{control.glmcat}
\alias{control.glmcat}
\title{control glmcat models}
\usage{
control.glmcat(maxit = 25, epsilon = 1e-06, beta_init = NA)
}
\arguments{
\item{maxit}{iterations}

\item{...}{other arguments.}
}
\description{
control LogLikelihood for GLMcat models.
}
