% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{control_glmcat}
\alias{control_glmcat}
\title{Control parameters for \code{glmcat} models}
\usage{
control_glmcat(maxit = 25, epsilon = 1e-06, beta_init = NA)
}
\arguments{
\item{maxit}{the maximum number of the Fisher's Scorng Algorithm iterations. Defaults to 25.}

\item{epsilon}{a double to change update the convergence criterion of GLMcat models.}

\item{beta_init}{an appropiate sized vector for the initial iteration of the algorithm.}
}
\description{
Set control parameters for \code{glmcat} models.
}
