% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcat2.R
\name{discrete_cm2}
\alias{discrete_cm2}
\title{Variance-Covariance Matrix for a Fitted glmcat Model Object}
\usage{
discrete_cm2(
  formula,
  case_id,
  alternatives,
  reference,
  alternative_specific,
  data,
  cdf = c("logistic", "normal", "gumbel", "gompertz", "cauchy", "laplace"),
  intercept = "standard",
  normalization = 1,
  control = list()
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit. An expression of the form y ~ predictors is interpreted as a specification that the response y is modelled by a linear predictor specified symbolically by model.}

\item{data}{a dataframe object in R, with the dependent variable as factor.}

\item{cdf}{\describe{
\item{\code{cdf}:}{a string indicating the F cdf, options are: logistic, normal, cauchy, student (any df), noncentralt, gompertz, gumbel and laplace.}
\item{\code{df}:}{an integer with the degrees of freedom of the 'cdf'}
\item{\code{mu}:}{an integer with the mu parameter of the 'cdf'}
}}

\item{normalization}{the quantile to use for the normalization of the estimated coefficients where the logistic distribution is used as the base cumulative distribution function.}

\item{control}{\describe{
\item{\code{maxit}:}{the maximum number of iterations for the Fisher scoring algorithm.}
\item{\code{epsilon}:}{a double to change update the convergence criterion of GLMcat models.}
\item{\code{beta_init}:}{an appropiate sized vector for the initial iteration of the algorithm.}
}}
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted \code{glmcat} model object.
}
