% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DisturbedDreamsDoc.R
\docType{data}
\name{DisturbedDreams}
\alias{DisturbedDreams}
\title{Severity of disturbed dreams}
\format{A dataframe containing :
\describe{
\item{Age}{Individuals age}
\item{Level}{Severity level: Not.severe, Severe.1, Severe.2, Very.severe.}
}}
\usage{
data(DisturbedDreams)
}
\description{
$Boy's disturbed dreams$ benchmark dataset drawn from a study that cross-classified boys by their age, and the severity (not severe, severe 1, severe 2, very severe) of their disturbed dreams (Maxwell, 1961).
}
\examples{
data(DisturbedDreams)
}
\references{
Maxwell, A.E. (1961) \emph{Analysing qualitative data}, Methuen London, 73.
}
\keyword{datasets}
