% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GLMcum}
\alias{GLMcum}
\title{Family of cumulative models for ordinal responses}
\usage{
GLMcum(
  formula,
  categories_order = NA_character_,
  proportional = NA_character_,
  data,
  distribution = "logistic",
  freedom_degrees = 1,
  beta_init = 1,
  threshold = "standard"
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit. An expression of the form y ~ model is interpreted as a specification that the response y is modelled by a linear predict_glmcator specified symbolically by model.}

\item{categories_order}{a character vector indicating the incremental order of the categories: c("a", "b", "c"); a<b<c. Alphabetical order is assumed by default.}

\item{proportional}{a character vector indicating the name of the variables with a proportional effect.}

\item{data}{a dataframe object in R, with the dependent variable as factor.}

\item{distribution}{an string indicating the F distribution, options are: logistic, normal, cauchit, student (any df), gompertz, gumbel.}

\item{freedom_degrees}{an optional scalar to indicate the degrees of freedom for the Student distribution.}

\item{beta_init}{optional beta initialization vector.}

\item{threshold}{restriction to impose on the the thresholds, options are: standard or equidistant.}
}
\value{
GLMcum returns a list which can be examined with the function summary.
}
\description{
Family of cumulative models for ordinal responses
}
\examples{
data(DisturbedDreams)
GLMcum(formula = Level ~ Age, threshold = "equidistant",
categories_order = c("Not.severe", "Severe.1", "Severe.2", "Very.severe"),
data = DisturbedDreams, distribution = "logistic")
}
