% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{logLik.glmcat}
\alias{logLik.glmcat}
\title{Log-likelihood of a fitted \code{glmcat} model object}
\usage{
\method{logLik}{glmcat}(object, ...)
}
\arguments{
\item{object}{an fitted object of class \code{glmcat}.}

\item{...}{additional arguments affecting the loglik.}
}
\description{
Extract Log-likelihood of a fitted \code{glmcat} model object.
}
