% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{nobs.glmcat}
\alias{nobs.glmcat}
\title{Number of observations in a glmcat model}
\usage{
\method{nobs}{glmcat}(object, ...)
}
\arguments{
\item{object}{a GLMcat model.}

\item{...}{other arguments.}
}
\description{
Extract the number of observations from a GLMcat model.
}
\examples{
data(DisturbedDreams)
mod1 <- GLMcat(
  formula = Level ~ Age,
  categories_order = c("Not.severe", "Severe.1", "Severe.2", "Very.severe"),
  data = DisturbedDreams, cdf = "logistic"
)
nobs(mod1)
}
