% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{noncentralt_glmcat}
\alias{noncentralt_glmcat}
\title{Noncentral t distribution for the \code{glmcat} models}
\usage{
noncentralt_glmcat(df = 7, mu = 0)
}
\arguments{
\item{df}{degrees of freedom of Student's t distribution.}

\item{mu}{the non centrality parameter.}
}
\description{
Noncentral t distribution to use as the inverse link of the \code{glmcat} models.
}
