% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{normalization.glmcat}
\alias{normalization.glmcat}
\title{normalization glmcat models}
\usage{
normalization.glmcat(p = 0.95, cdf = "logistic", degrees_freedom = 7)
}
\arguments{
\item{p}{percentile}

\item{cdf}{cdf}

\item{degrees_freedom}{cdf}
}
\description{
normalization of coefficients of GLMcat models.
}
