% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{plot.glmcat}
\alias{plot.glmcat}
\title{Plot method for a fitted code{glmcat} model object}
\usage{
\method{plot}{glmcat}(x, ...)
}
\arguments{
\item{x}{an object of class \code{glmcat}.}

\item{...}{additional arguments.}
}
\description{
\code{plot} of the log-likelihood profile for a fitted \code{glmcat} model object.
}
