% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R, R/RcppExports.R
\name{predict.glmcat}
\alias{predict.glmcat}
\alias{predict_glmcat}
\title{Predict Method for a Fitted glmcat Model Object}
\usage{
\method{predict}{glmcat}(object, newdata, type, ...)

predict_glmcat(model_object = NaN, data, type = "prob")
}
\arguments{
\item{object}{an fitted object of class \code{"glmcat"}.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{type}{The type of prediction to obtain. \code{"prob"} gives probabilities
and \code{"linear.predict"} gives the linear predictor.}

\item{model_object}{a GLMcat model}

\item{data}{a data frame with the predictor variables used in the GLMcat model.}

\item{...}{additional arguments affecting the predict produced.}
}
\description{
Obtains predictions from a fitted \code{"glmcat"} object.

GLMcat model predictions
}
\details{
Prediction based on GLMcat models
}
\examples{
data(DisturbedDreams)
mod1 <- GLMcat(formula = Level ~ Age,
data = DisturbedDreams, cdf = "logistic")
predict_glmcat(mod1, data = DisturbedDreams[1:5, ], type = "prob")
}
