% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{print.anova.glmcat}
\alias{print.anova.glmcat}
\title{Printing Anova for \code{glmcat} model fits}
\usage{
\method{print}{anova.glmcat}(
  x,
  digits = max(getOption("digits") - 2, 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{an object of class \code{"glmcat"}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{print.anova} method for GLMcat objects.
}
