% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{print.glmcat}
\alias{print.glmcat}
\title{Print method for an "glmcat" object.}
\usage{
\method{print}{glmcat}(x, ...)
}
\arguments{
\item{...}{additional arguments affecting the summary produced.}

\item{object}{an object of class \code{"glmcat"}.}
}
\description{
\code{print} method for a fitted \code{glmcat} model object.
}
