% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.summary.glmcat}
\alias{print.summary.glmcat}
\title{Printing a fitted \code{glmcat} model object}
\usage{
\method{print}{summary.glmcat}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{an object of class \code{"glmcat"}.}

\item{digits}{the number of digits in the printed table.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{print.summary} method for GLMcat objects.
}
