% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.summary.glmcat}
\alias{print.summary.glmcat}
\title{Printing a fitted \code{glmcat} model object}
\usage{
\method{print}{summary.glmcat}(
  x,
  digits = max(3, getOption("digits") - 3),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{object}{an object of class \code{"glmcat"}.}
}
\description{
\code{print.summary} method for GLMcat objects.
}
