% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{safe_pchisq}
\alias{safe_pchisq}
\title{#' Student's t distribution for the \code{glmcat} models
#' @description Student's t distribution to use as the inverse link of the \code{glmcat} models.
#' @rdname student_glmcat
#' @param df degrees of freedom of Student's t distribution.
#' @export
student_glmcat <- function(df = 7) {
  return(list("cdf" = "student", "df" = df))
}}
\usage{
safe_pchisq(q, df, ...)
}
\description{
#' Noncentral t distribution for the \code{glmcat} models
#' @description Noncentral t distribution to use as the inverse link of the \code{glmcat} models.
#' @rdname noncentralt_glmcat
#' @param df degrees of freedom of Student's t distribution.
#' @param mu the non centrality parameter.
#' @export
noncentralt_glmcat <- function(df = 7, mu = 0) {
  return(list("cdf" = "noncentralt", "df" = df, "mu" = mu))
}
}
