% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{step_glmcat}
\alias{step_glmcat}
\title{Stepwise for glmcat models}
\usage{
step_glmcat(
  object,
  scope,
  scale = 0,
  direction = c("both", "backward", "forward"),
  trace = 1,
  keep = NULL,
  steps = 1000,
  k = 2,
  ...
)
}
\description{
Stepwise
}
