% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{student_glmcat}
\alias{student_glmcat}
\title{Student's t distribution for the \code{glmcat} models}
\usage{
student_glmcat(df = 7)
}
\arguments{
\item{df}{degrees of freedom of Student's t distribution.}
}
\description{
Student's t distribution to use as the inverse link of the \code{glmcat} models.
}
