% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{terms.glmcat}
\alias{terms.glmcat}
\title{Terms of a fitted \code{glmcat} model object}
\usage{
\method{terms}{glmcat}(x, ...)
}
\arguments{
\item{x}{an object of class \code{glmcat}.}

\item{...}{additional arguments.}
}
\description{
Returns the terms of a fitted \code{glmcat} model object.
}
