% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.glmcat}
\alias{anova.glmcat}
\title{Anova for a fitted \code{glmcat} model object}
\usage{
\method{anova}{glmcat}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"glmcat"}.}
}
\description{
Compute an analysis of deviance table for one fitted \code{glmcat} model object.
}
