% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GLMcat}
\alias{GLMcat}
\title{Family of models for categorical responses}
\usage{
GLMcat(
  formula,
  ratio = "reference",
  distribution = "logistic",
  categories_order,
  proportional_effects = NA_character_,
  data,
  freedom_degrees = 1L
)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit. An expression of the form y ~ model is interpreted as a specification that the response y is modelled by a linear predict_glmcator specified symbolically by model.}

\item{ratio}{an string indicating the F distribution, options are: reference, adjacent, cumulative and sequential.}

\item{distribution}{an string indicating the F distribution, options are: logistic, normal, cauchit, student (any df), gompertz, gumbel.}

\item{categories_order}{a character vector indicating the incremental order of the categories: c("a", "b", "c"); a<b<c}

\item{proportional_effects}{a character vector indicating the name of the variables with a proportional effect.}

\item{data}{a dataframe object in R, with the dependent variable as factor.}

\item{freedom_degrees}{an optional scalar to indicate the degrees of freedom for the Student distribution.}
}
\value{
GLMcat returns a list which can be examined with the function summary.
}
\description{
Family of models for categorical responses
}
\examples{
library(GLMcat)
data(DisturbedDreams)
GLMcat(formula = Level ~ Age,
categories_order = c("Not.severe", "Severe.1", "Severe.2", "Very.severe"),
data = DisturbedDreams, distribution = "logistic", ratio = "reference")
}
