% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{logLik.glmcat}
\alias{logLik.glmcat}
\title{Log-likelihood glmcat models}
\usage{
\method{logLik}{glmcat}(object, ...)
}
\arguments{
\item{object}{a GLMcat model.}

\item{...}{other arguments.}
}
\description{
Extract Log-likelihood for glmcat models.
}
\examples{
data(DisturbedDreams)
mod1 <- GLMcat(formula = Level ~ Age,
categories_order = c("Not.severe", "Severe.1", "Severe.2", "Very.severe"),
data = DisturbedDreams, distribution = "logistic")
logLik(mod1)
}
