% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{nobs.glmcat}
\alias{nobs.glmcat}
\title{Number of observations of a fitted \code{glmcat} model object}
\usage{
\method{nobs}{glmcat}(object, ...)
}
\arguments{
\item{object}{an fitted object of class \code{glmcat}.}

\item{...}{additional arguments affecting the \code{nobs} method.}
}
\description{
Extract the number of observations of the fitted \code{glmcat} model object.
}
