% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{noncentralt.glmcat}
\alias{noncentralt.glmcat}
\title{Noncentral t cdf for glmcat models}
\usage{
noncentralt.glmcat(df = 7, mu = 0)
}
\arguments{
\item{df}{degrees_freedom}

\item{mu}{non centrality parameter}
}
\description{
Noncentral t cdf
}
