% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{predict.glmcat}
\alias{predict.glmcat}
\title{Predict Method for a Fitted glmcat Model Object}
\usage{
\method{predict}{glmcat}(object, newdata, type, ...)
}
\arguments{
\item{object}{an fitted object of class \code{"glmcat"}.}

\item{newdata}{optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{type}{the type of prediction required. The default is on the scale of the linear predictors; ' \code{"prob"} gives probabilities and \code{"linear.predictor"} gives predictions on the scale of thelinear predictor including the boundary categories.}

\item{...}{additional arguments affecting the predict produced.}
}
\description{
Obtains predictions from a fitted \code{"glmcat"} object.
}
