% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_glmcat}
\alias{predict_glmcat}
\title{Prediction based on GLMcat models}
\usage{
predict_glmcat(model_object = NaN, data, type = "prob")
}
\arguments{
\item{model_object}{a GLMcat model}

\item{data}{a data frame with the predictor variables used in the GLMcat model.}

\item{type}{The type of prediction to obtain. \code{"prob"} gives probabilities
and \code{"linear.predict"} gives the linear predictor.}
}
\description{
GLMcat model predictions
}
