% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{print.glmcat}
\alias{print.glmcat}
\title{Print method for a fitted code{glmcat} model object}
\usage{
\method{print}{glmcat}(x, ...)
}
\arguments{
\item{x}{an object of class \code{glmcat}.}

\item{...}{additional arguments.}
}
\description{
\code{print} method for a fitted \code{glmcat} model object.
}
