% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{print.glmcat}
\alias{print.glmcat}
\title{Print method for a fitted code{glmcat} model object.}
\usage{
\method{print}{glmcat}(object, ...)
}
\arguments{
\item{object}{an object of class \code{glmcat}.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{print} method for a fitted \code{glmcat} model object.
}
