% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R, R/step.R
\name{step.glmcat}
\alias{step.glmcat}
\title{Stepwise for a \code{glmcat} model object}
\usage{
\method{step}{glmcat}(object, scope, scale, direction, trace, keep, steps, k, ...)
}
\arguments{
\item{object}{an fitted object of class \code{glmcat}.}

\item{scope}{defines the range of models examined in the stepwise search (same as in the step function of the stats package). This should be either a single formula, or a list containing components upper and lower, both formulae.}

\item{scale}{not used in this function; included for compatibility.}

\item{direction}{the mode of the stepwise search.}

\item{trace}{to print the process information.XLASL}

\item{keep}{not used in this function; included for compatibility.}

\item{steps}{the maximum number of steps.}

\item{k}{not used in this function; included for compatibility.}

\item{...}{additional arguments.}
}
\description{
Stepwise for a \code{glmcat} model object based on the AIC.

}
