% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{step_glmcat}
\alias{step_glmcat}
\title{Stepwise for glmcat models}
\usage{
step_glmcat(
  object,
  data,
  scope,
  direction = c("both", "backward", "forward"),
  trace = 1,
  steps = 1000
)
}
\arguments{
\item{object}{a GLMcat model.}

\item{scope}{defines the range of models examined in the stepwise search (same as in the step function of the stats package). This should be either a single formula, or a list containing components upper and lower, both formulae.}

\item{direction}{the mode of the stepwise search.}

\item{trace}{to print the process information.}

\item{steps}{the maximum number of steps.}
}
\description{
Stepwise based on the AIC
}
