% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{summary.glmcat}
\alias{summary.glmcat}
\title{Summary of models}
\usage{
\method{summary}{glmcat}(object, ...)
}
\arguments{
\item{object}{a GLMcat model}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
\code{summary} method for GLMcat objects.
}
