% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{summary.glmcat}
\alias{summary.glmcat}
\title{Summarising \code{glmcat} Model Fits}
\usage{
\method{summary}{glmcat}(object, correlation = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{"glmcat"}.}
}
\description{
\code{summary} method for GLMcat objects.
}
