% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMcatMethods.R
\name{terms.glmcat}
\alias{terms.glmcat}
\title{Terms of a fitted \code{glmcat} model object}
\usage{
\method{terms}{glmcat}(object)
}
\arguments{
\item{object}{an object of class \code{glmcat}.}
}
\description{
Returns the terms of a fitted \code{glmcat} model object.
}
